setup.sh: GECOS for alp; disallow passwd login; create ~alp/.profile
The .profile is me trying to figure out how /bin/ash does its config for
an interactive shell.  I probably should give up and use bash?

Disallowing password login is to see if that lets alp login with an ssh
key.  Right now alp's password is blank ("-D") option to `adduser`;  alp
ssh'ing fails with the logged message (on the server):

User alp not allowed because account is locked

On the client side, it reports:

Received disconnect from IP port 22:2: Too many authentication failures

Which... there *are* a lot of failed ssh logins (on a brand new box!
wow), but... I think it might just be that alp has a blank password.

Filling in the GECOS field is the tiniest of QOL details.